add relaydb
